% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{fit_hierarchical_analytical}
\alias{fit_hierarchical_analytical}
\title{Analytical shrinkage backend}
\usage{
fit_hierarchical_analytical(data, prior_spec, verbose = TRUE)
}
\arguments{
\item{data}{Data frame with columns: zone (factor), ilr1, ilr2, ... (ILR values)}

\item{prior_spec}{An object of class \code{"gc_prior_spec"}}

\item{verbose}{Logical, print progress (default TRUE)}
}
\value{
S3 object of class \code{"gc_hierarchical_fit"} with analytical results
}
\description{
Fit hierarchical model using empirical Bayes analytical shrinkage.
Provides fast closed-form estimates suitable for exploration and visualization.
}
\keyword{internal}
