% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sda_integration.R
\name{gc_fetch_mapunit_polygons}
\alias{gc_fetch_mapunit_polygons}
\title{Fetch Map Unit Polygons as Spatial Zones}
\usage{
gc_fetch_mapunit_polygons(extent, crs = 4326, verbose = TRUE)
}
\arguments{
\item{extent}{Spatial object (sf, raster, bbox) defining geographic extent}

\item{crs}{Coordinate reference system (default EPSG:4326)}

\item{verbose}{Logical, print query progress (default TRUE)}
}
\value{
An sf object with MUPOLYGON geometries and columns:
\itemize{
\item \code{mukey}: Map unit key (SSURGO ID)
\item \code{musym}: Map unit symbol
\item \code{muname}: Map unit name
\item \code{geometry}: sf polygons
}
}
\description{
Query Soil Data Access (SDA) to retrieve map unit polygons (MUPOLYGON)
within a geographic extent. Polygons serve as spatial zones for zone-stratified
geostatistical modeling.
}
\details{
Uses \code{soilDB::SDA_spatialQuery(what="mupolygon", geomIntersection=TRUE)}
to retrieve map unit boundaries intersecting the extent.
}
\examples{
\dontrun{
library(sf)
extent <- st_bbox(c(xmin = -93.5, ymin = 41.5,
                    xmax = -93.4, ymax = 41.6), crs = 4326)

mapunit_zones <- gc_fetch_mapunit_polygons(extent)
print(nrow(mapunit_zones))
}

}
\seealso{
\code{\link[soilDB:SDA_spatialQuery]{soilDB::SDA_spatialQuery()}}
}
