% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_discretization.R
\name{gc_standardize_depths_linear}
\alias{gc_standardize_depths_linear}
\title{Internal: Linear interpolation depth standardization}
\usage{
gc_standardize_depths_linear(prof, intervals, comp_cols, id_col, prof_id)
}
\description{
Internal: Linear interpolation depth standardization
}
\keyword{internal}
