% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_discretization.R
\name{gc_standardize_depths_spline}
\alias{gc_standardize_depths_spline}
\title{Internal: Spline-based depth standardization}
\usage{
gc_standardize_depths_spline(prof, intervals, comp_cols, id_col, prof_id)
}
\description{
Internal: Spline-based depth standardization
}
\keyword{internal}
